financial-reports-generator/requirements.txt
pandas
streamlit
gradio
openai
markdown
faker
python-dotenv
tabulate
plotly
matplotlib
scikit-learn